requirements.txt
# requirements.txt

asgiref==3.8.1
attrs==23.2.0
bleach==6.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-api-logger==1.1.16
drf-spectacular==0.27.2
drf-yasg==1.21.7
idna==3.7
importlib_resources==6.4.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
psycopg2-binary==2.9.9
Pygments==2.18.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.0
setuptools==71.1.0
six==1.16.0
sqlparse==0.5.0
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
webencodings==0.5.1
